+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased][unreleased]
 
+### Fixed
+- fix set_config null handling: empty string "" value can now be saved (fixes #505)
+
 ## [1.56.0] - 2021-27-06
 
 ### Changed